Add a class to the repository from the sheet.
This is currently limited to NSObject subclasses, because the UI does not

give you any way to choose a superclass. I don't want to make that a free form

field, so I'll add a 'create subclass' action that can be done on existing

classes that uses the same UI as the one created here.